Convert hidden link to nowhere to div
I don't understand the signifigance of this link. It doesn't go
anwyhere, and the svg in it is hidden. Because it references an
undefined anchor, it causes htmlproof to complain. However, it is referenced in the
SASS and removing it breaks the layout. Converting it to a div seems to
preserve the page formatting and make htmlproof happy.