Replaced slack form with slackin link
slackin (http://rauchg.com/slackin/) is an automatic Slack signup form that lets user's join the Slack page with a single click. I'm hosting an endpoint on my Heroku account at https://nashville-brigade-slackin.herokuapp.com/. I replaced the link to the Slack form with a link to this new endpoint.

This endpoint also exposes badges we can post for individual channels we can add to README of projects (a la Travis CI badges). Haven't done any work on it yet, but it could be a cool addition and help users get involved in the discussion more quickly.